CircularCloudLayouter.PlaceNewRectangle() [WIP]
