+++ cs/TagsCloudVisualization/TagsCloudVisualization/Program.cs
@@ -9,7 +9,6 @@ namespace TagsCloudVisualization
     {
         static void Main(string[] args)
         {
-            Console.WriteLine((int)1.99999999999);
 
         }
     }
